+++ src/Exceptions/ApiError.php
@@ -5,11 +5,14 @@ declare(strict_types=1);
 namespace Paddle\SDK\Exceptions;
 
 use Psr\Http\Client\ClientExceptionInterface;
+use Psr\Http\Message\ResponseInterface;
 
 class ApiError extends \Exception implements ClientExceptionInterface
 {
     /** @var array<FieldError> */
     public array $fieldErrors;
+
+    public readonly int|null $retryAfter;
 
     final public function __construct(
         public string $type,
@@ -23,9 +26,9 @@ class ApiError extends \Exception implements ClientExceptionInterface
         parent::__construct($this->detail);
     }
 
-    public static function fromErrorData(array $error): static
+    public static function fromErrorData(array $error, ResponseInterface|null $response = null): static
     {
-        return new static(
+        $apiError = new static(
             $error['type'],
             $error['code'],
             $error['detail'],
@@ -35,5 +38,10 @@ class ApiError extends \Exception implements ClientExceptionInterface
                 $error['errors'] ?? [],
             ),
         );
+
+        $retryAfterHeader = $response?->getHeader('Retry-After')[0] ?? null;
+        $apiError->retryAfter = $retryAfterHeader ? intval($retryAfterHeader) : null;
+
+        return $apiError;
     }
 }